backend/requirements.txt

pytest==9.0.2
SQLAlchemy==2.0.48
fastapi[standard]
python-dateutil==2.9.0.post0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
uvicorn==0.34.0
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
weasyprint==64.0
fastapi-mail==1.4.1
jinja2==3.1.5
pytz==2024.1
alembic==1.18.4
openpyxl==3.1.5
XlsxWriter==3.2.2
